Add ?next= to edit url to redirect back to the right tree after edit.
At the moment the user returns to explore view of the parent page, when
the non-canonical tree is hidden, this doesn't display the child pages,
which makes it hard to edit other child pages.

By adding ``?next=/url/to/parent-canonical`` the user gets redirected to
to canonical page of the parent, which is the expecting behavior and
allows the user to edit pages again in different languages with the
``edit in`` dropdown.